Create Tool Struct
I decided to make the Tool class into a struct instead. The reason for this
change is that I couldn't find any methods that the tool would actually
use. Since the simulation doesn't actually have the customer 'use' the tool
it just tracks who has it and when it needs to be returned. So in my design
I've decided to change Tool from a class to a struct.